requirements.txt
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
Pillow==10.4.0
streamlit==1.36.0
numpy==1.26.4